requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# Data sources
yfinance>=0.2.36

# Configuration
pyyaml>=6.0

# Web dashboard
streamlit>=1.30.0
plotly>=5.18.0

# Logging
python-json-logger>=2.0.0

# Validation
jsonschema>=4.20.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
